refact: move utils functions to utils folder and add common config utils files there
