+++ packages/stg-evm-v2/devtools/config/mainnet/01/credit-messaging.config.ts
@@ -3,13 +3,12 @@ import { CreditMessagingEdgeConfig, CreditMessagingNodeConfig } from '@stargatef
 import { OmniGraphHardhat } from '@layerzerolabs/devtools-evm-hardhat'
 
 import { generateCreditMessagingConfig } from '../../utils'
-
-import buildMessagingGraph from './messaging.config.utils'
+import { buildMessagingGraphMainnet } from '../utils'
 
 const contract = { contractName: 'CreditMessaging' }
 
 export default async (): Promise<OmniGraphHardhat<CreditMessagingNodeConfig, CreditMessagingEdgeConfig>> => {
-    return buildMessagingGraph(contract, 'CREDIT_MESSAGING', generateCreditMessagingConfig) as Promise<
+    return buildMessagingGraphMainnet(contract, 'CREDIT_MESSAGING', generateCreditMessagingConfig) as Promise<
         OmniGraphHardhat<CreditMessagingNodeConfig, CreditMessagingEdgeConfig>
     >
 }
