+++ packages/stg-evm-v2/devtools/config/mainnet/01/feelib-v1.meth.config.ts
@@ -3,10 +3,10 @@ import { FeeLibV1EdgeConfig, FeeLibV1NodeConfig } from '@stargatefinance/stg-dev
 
 import { OmniGraphHardhat } from '@layerzerolabs/devtools-evm-hardhat'
 
-import buildFeeLibV1DeploymentGraph from './feelib-v1.config.utils'
+import { buildFeeLibV1DeploymentGraphMainnet } from '../utils'
 
 const tokenName = TokenName.mETH
 
 export default async (): Promise<OmniGraphHardhat<FeeLibV1NodeConfig, FeeLibV1EdgeConfig>> => {
-    return buildFeeLibV1DeploymentGraph(tokenName)
+    return buildFeeLibV1DeploymentGraphMainnet(tokenName)
 }
