+++ packages/stg-evm-v2/devtools/config/mainnet/01/oft-token.config.ts
@@ -6,7 +6,7 @@ import { OmniGraphHardhat, createGetHreByEid } from '@layerzerolabs/devtools-evm
 import { getTokenDeployName, getUSDTDeployName } from '../../../../ops/util'
 import { createGetAssetAddresses, getAssetType } from '../../../../ts-src/utils/util'
 import { getContractWithEid, getSafeAddress } from '../../utils'
-import { getChainsThatSupportTokenWithType, getChainsThatSupportsUsdtOftByDeployment } from '../../utils.config'
+import { getChainsThatSupportTokenWithType, getChainsThatSupportsUsdtOftByDeployment } from '../../utils/utils.config'
 import { setMainnetStage } from '../utils'
 
 // Both USDC and USDT now (as of 2024-12-10) have their own config files, so this file is just used for WETH Hydra deployments
