+++ packages/stg-evm-v2/devtools/config/utils/feelib-v1.config.utils.ts
@@ -2,19 +2,20 @@ import { TokenName } from '@stargatefinance/stg-definitions-v2'
 import { FeeLibV1EdgeConfig, FeeLibV1NodeConfig } from '@stargatefinance/stg-devtools-v2'
 
 import { OmniGraphHardhat } from '@layerzerolabs/devtools-evm-hardhat'
+import { Stage } from '@layerzerolabs/lz-definitions'
 
-import { getFeeLibV1DeployName } from '../../../../ops/util'
-import { getContractWithEid } from '../../utils'
-import { filterValidProvidedChains, getChainsThatSupportToken, printChains } from '../../utils.config'
-import { setMainnetStage } from '../utils'
+import { getFeeLibV1DeployName } from '../../../ops/util'
+import { getContractWithEid } from '../utils'
 
-import { DEFAULT_PLANNER } from './constants'
+import { filterValidProvidedChains, getChainsThatSupportToken, printChains, setStage } from './utils.config'
 
 export default async function buildFeeLibV1DeploymentGraph(
-    tokenName: TokenName
+    stage: Stage,
+    tokenName: TokenName,
+    defaultPlanner: string
 ): Promise<OmniGraphHardhat<FeeLibV1NodeConfig, FeeLibV1EdgeConfig>> {
-    // Set the stage to mainnet
-    setMainnetStage()
+    // Set the correct stage
+    setStage(stage)
 
     const contract = { contractName: getFeeLibV1DeployName(tokenName) }
 
@@ -22,7 +23,7 @@ export default async function buildFeeLibV1DeploymentGraph(
     const chainsList = process.env.CHAINS_LIST ? process.env.CHAINS_LIST.split(',') : []
 
     const defaultNodeConfig = {
-        owner: DEFAULT_PLANNER,
+        owner: defaultPlanner,
     }
 
     // get valid chains config in the chainsList
