+++ packages/stg-evm-v2/test/devtools/asset.config.test.ts
@@ -14,7 +14,7 @@ import assetUsdcConfig from '../../devtools/config/mainnet/01/asset.usdc.config'
 import assetUsdtConfig from '../../devtools/config/mainnet/01/asset.usdt.config'
 import { DEFAULT_PLANNER } from '../../devtools/config/mainnet/01/constants'
 import { filterConnections, generateAssetConfig } from '../../devtools/config/utils'
-import { getAllSupportedChains, getChainsThatSupportToken } from '../../devtools/config/utils.config'
+import { getAllSupportedChains, getChainsThatSupportToken } from '../../devtools/config/utils/utils.config'
 
 import { setupConfigTestEnvironment } from './utils'
