+++ packages/stg-evm-v2/test/devtools/rewarder.config.test.ts
@@ -5,7 +5,7 @@ import hre from 'hardhat'
 
 import rewarderConfig from '../../devtools/config/mainnet/01/rewarder.config'
 import { getSafeAddress } from '../../devtools/config/utils'
-import { getAllSupportedChains, getChainsThatSupportRewarder } from '../../devtools/config/utils.config'
+import { getAllSupportedChains, getChainsThatSupportRewarder } from '../../devtools/config/utils/utils.config'
 
 import { setupConfigTestEnvironment } from './utils'
